Rework logic for calculating digital take-up
If the number of normal transactions is either not provided or is 0,
we're unable to calculate digital take-up so return 'no data'.

Otherwise, the number of transactions is a positive integer,
so if the number of digital transactions is 0 then digital take-up
is 0.

Otherwise, if the number of digital transactions is not provided
then provide 'no data'.

Otherwise do the calculation.