Add create csv script
The transaction-volumes.csv that is linked to from transactions
explorer is generated everytime the services dashboards are
generated as part of the create pages script.

It has been extracted from create pages to reside with the rest
of the spreadsheet functions because:
1. Some users may still need the services data in this format, and
previously there was no way to generate the csv file without
recreating the transactions explorer site.
2. Transactions explorer will soon be archived.

Update the version of python travis runs against.

The script to generate the transaction volumes data csv uses a
built in python function that doesn't exist in version 2.6

Changed the travis config to only run the tests against python 2.7.