Fix definitions of year and pep8
I do not agree with the existing definition of year. The end at of a year
should be the day after the final day. The start at should be the start
of the first quarter. This is the pattern we take with quarterly data -
years should be consistent with this. This will require clearing out the existing
datasets but I think it's worth it for clarity. I might have missed
something with this though but I can't find anything in the commit
messages or comments.