This is what the spreadsheet implies
it's value for jan to dec 2014 are the same as it's rolling value for q1
2015. If the jan to dec number is right then the spreadsheet q1 2015
should become q4 2014. if the q1 2015 number is right then the jan to
dec number should be updated to the value for q4 2014. Alternatively the
timestamps in the process file should change as here or as in fix years.