Replace Google sheets handling with simple csv.
Lots of workflows through this codebase no longer work and given the
attitude towards the product in general at the moment, and the increased
likelihood a developer / engineer is going to be running this next time,
we have decided to replace the complex Google sheets handling code with
a simpler local csv version.

A diff of the generated JSON was checked to make sure the behaviour was
the same.