Split JSON payload into multiple, smaller chunks
In production we saw 503 errors with the message:

> GOV.UK is experiencing technical difficulties now. Please try again later.

This appears to originate from Fastly[1] which sits in front of the
Performance Platform in production. After a little bit of poking it
appears that the payload we were sending was too large (~14MB).

We split up the payload and send it as multiple, smaller chunks to work
around this size limit.

[1] https://github.com/alphagov/govuk-cdn-config/blob/9aa67ac74286f739bb97bf9e53854458f88f1a4f/vcl_templates/performanceplatform.vcl.erb